alterações na main

+++ src/ContaSalario.java
@@ -1,14 +1,22 @@
+import java.util.ArrayList;
+
 public class ContaSalario extends Conta{
     private double limSaque, limTransacao;
 
-    public ContaSalario(int senha, int nroConta,Agencia ag,float saldo,Cliente cliente,double limTransacao,double limSaque)
+    public ContaSalario(int senha, int nroConta,Agencia ag,double saldo,Cliente cliente,double limTransacao,double limSaque)
     {
         super(senha,nroConta,ag,saldo,cliente);
-
         this.limSaque = limSaque;
         this.limTransacao = limTransacao;
-
     }
+
+//    // construtor para mais de um cliente
+//    public ContaSalario(int senha, int nroConta, Agencia ag, double saldo, ArrayList<Cliente> clientes, double limTransacao, double limSaque)
+//    {
+//        super(senha,nroConta,ag,saldo,clientes);
+//        this.limSaque = limSaque;
+//        this.limTransacao = limTransacao;
+//    }
 
     @Override
     public void deposito(float val, String canal) throws IllegalArgumentException {
